requirements.txt
torch
pillow
transformers

openai
hydra-core
pytz
wandb
tenacity